Update VPN script
added space for if command in rancid functions so it is recognised in IDE

added function to run rancid config update after actions taken on the device to record it as it happens.

There was a elif on line 215 that looked for the existence of a variable that was not defined. I have changed this to the variable that it deletes if found. Will test and see whether this fixes VPN Fails #4